Split utility code from sampler for easier reuse
