+++ lib/sampler.js
@@ -12,58 +12,13 @@
 const Stream = require('stream');
 
 const ThingTalk = require('thingtalk');
-const Type = ThingTalk.Type;
 
 const { coin, uniform, choose } = require('./random');
+const { entityTypeToTTType, makeLookupKeys } = require('./sample-utils');
 const i18n = require('./i18n');
 
 class UnassignableEntity extends Error {}
 class SkippedEntity extends Error {}
-
-const TYPES = {
-    QUOTED_STRING: Type.String,
-    NUMBER: Type.Number,
-    CURRENCY: Type.Currency,
-    DURATION: Type.Measure('ms'),
-    LOCATION: Type.Location,
-    DATE: Type.Date,
-    TIME: Type.Time,
-
-    EMAIL_ADDRESS: Type.Entity('tt:email_address'),
-    PHONE_NUMBER: Type.Entity('tt:phone_number'),
-    HASHTAG: Type.Entity('tt:hashtag'),
-    USERNAME: Type.Entity('tt:username'),
-    URL: Type.Entity('tt:url'),
-    PATH_NAME: Type.Entity('tt:path_name'),
-};
-
-function entityTypeToTTType(entityType, unit) {
-    if (entityType === 'NUMBER' && !!unit)
-        return Type.Measure(unit);
-    else if (entityType.startsWith('GENERIC_ENTITY_'))
-        return Type.Entity(entityType.substring('GENERIC_ENTITY_'.length));
-    else if (entityType.startsWith('MEASURE_'))
-        return Type.Measure(entityType.substring('MEASURE_'.length));
-    else
-        return TYPES[entityType];
-}
-
-function makeLookupKeys(deviceFunctionName, param, type) {
-    const keys = [];
-    keys.push(String(type));
-    if (param)
-        keys.push(`param:${param}:${type}`);
-    if (param && deviceFunctionName) {
-        const dot = deviceFunctionName.lastIndexOf('.');
-        const deviceName = deviceFunctionName.substring(0, dot);
-        const functionName = deviceFunctionName.substring(dot+1);
-
-        keys.push(`param:${deviceName}.*:${param}:${type}`);
-        keys.push(`param:${deviceName}.${functionName}:${param}:${type}`);
-    }
-    keys.reverse();
-    return keys;
-}
 
 class SentenceProcessor {
     constructor(schemaRetriever, constants, detokenizer, options, input) {
